Change being made as of 8-4-2021:
To start: currently, elo is calculated by taking the difference between your rating and your performance rating (Delta P), times your volatility (determined by amount of matches played), times a scaling factor that exists to make larger changes depending on the expected score of your team. This third term, the "blowout factor" is currently being calculated as

blowoutTerm = abs( score - expectedScore ) / (2 * n_chapters)

 resulting in the black curve. This term punishes 5-0's extra hard by scaling the change in elo higher the harder your team won/lost form your expected score. Expected results are scaled down, seeing as players performed as expected. This is a good thing, as unexpected results should be awarded or punished, but seeing as the expected and eventual scores are also used in calculating your Delta P you do get extreme results at the extreme ends of the spectrum.

I suggest we modify the system to reward a more consistent performance and not be defined by your largest losses/wins by limiting the effect of this term: in extreme cases, the value of the blowout term would be limited to a 60% deviation of your expected score, resulting in less extreme changes with blowouts and more consistent changes on small deviations from your rating.

The new formula would be:

blowoutTerm = 0.04 * (score - expectedScore) ^2 + 0.03
with a limit on 60% deviation, resulting in the combination of red lines. With scores around expected we follow the parabola, outside of those we follow the flat line.

Included ability to reenter prev results by entering "P" on main menu.